% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagramsToGenerate.R
\name{diagramsToGenerate}
\alias{diagramsToGenerate}
\title{Title}
\usage{
diagramsToGenerate(elems)
}
\arguments{
\item{elems}{list of elements for which the diagrams shall be generated}
}
\value{
data.frame with the diagrams to be generated
}
\description{
Title
}
