% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodisctoponator-package.R
\docType{package}
\name{geodisctoponator-package}
\alias{geodisctoponator}
\alias{geodisctoponator-package}
\title{geodisctoponator: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Gerald Schuberth-Hlavač \email{abuseki@synapticgap.com}

}
\keyword{internal}
