% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverlapInfo.R
\name{getOverlapInfo}
\alias{getOverlapInfo}
\title{Title}
\usage{
getOverlapInfo(regions)
}
\arguments{
\item{regions}{named list of convex hull coordinates. Names are the
categories and their convex hulls coordinates are presented in x-y-matrix
like datastructures. This is what \code{\link[=processDiagram]{processDiagram()}} returns as
\code{catReducedHulls}}
}
\value{
List of
\itemize{
\item \code{olAreas}: The overlapping areas of regions
\item \code{olRates}: The overlapping rates per region
\item \code{regionsAreas}: Areas of single regions
}
}
\description{
Title
}
\seealso{
\code{\link[=processDiagram]{processDiagram()}}
}
