% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{getPlotForDiagram}
\alias{getPlotForDiagram}
\title{Title}
\usage{
getPlotForDiagram(
  dataSrc,
  xel = "x",
  yel = "y",
  categoryCol = "category",
  scaleAxes = TRUE,
  facets = FALSE
)
}
\arguments{
\item{facets}{}
}
\description{
Title
}
