% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducePointSet.R
\name{getPointsToRemove}
\alias{getPointsToRemove}
\title{Title}
\usage{
getPointsToRemove(ps, confCoef = 0.85, reduRat = 0.95, adjuFact = 0.05)
}
\arguments{
\item{confCoef}{confidence coefficient}

\item{reduRat}{reduction ratio}

\item{adjuFact}{adjustment factor of the reduction ratio}
}
\description{
Reduces a point set to a given confidence coefficient using Delaunay
triangulation as suggested by \insertCite{Han_2020;textual}{didigenr}
}
\references{
{
\insertRef{Han_2020}{didigenr}
}
}
\seealso{
\code{\link[=processDiagram]{processDiagram()}}
}
