% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getWorkingData}
\alias{getWorkingData}
\title{Title}
\usage{
getWorkingData(dataSrc, xel = "x", yel = "y", categoryCol = "category")
}
\arguments{
\item{dataSrc}{Queried data frame to get the elements or their ratios}

\item{xel}{Element or element ratio on the diagram's x axis}

\item{yel}{Element or element ratio on the diagram's y axis}

\item{categoryCol}{Name of the column in \code{dataSrc} holding the categories}
}
\description{
Normalized data set
}
