% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagramsToGenerate.R
\name{nDiagsForElems}
\alias{nDiagsForElems}
\title{Title}
\usage{
nDiagsForElems(nElems)
}
\arguments{
\item{nElems}{number of elements for which the number of diagrams will be
calcuated}
}
\description{
Simple function to calculate the number of element (element ratio) pairs for
a given number of elements. This is also the number of diagrams that will
have to be processed.
}
