% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_overlapRates.R
\name{overlapRates}
\alias{overlapRates}
\title{Title}
\usage{
overlapRates(regions)
}
\arguments{
\item{regions}{named list of regions as polygon coordinates. Coordinates must
be named \code{x} and \code{y}}
}
\value{
data.frame with the calculated overlap rates of the given polygons
}
\description{
TODO put formulaes here
}
