% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plotDiagram}
\alias{plotDiagram}
\title{Title}
\usage{
plotDiagram(
  dataSrc,
  xel = "x",
  yel = "y",
  categoryCol = "category",
  logCoords = FALSE,
  facets = FALSE,
  ...
)
}
\arguments{
\item{facets}{}
}
\description{
Title
}
