% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plotForDiagramData}
\alias{plotForDiagramData}
\title{Title}
\usage{
plotForDiagramData(
  dd,
  facets = FALSE,
  showPoints = TRUE,
  showHulls = TRUE,
  showDensity = FALSE,
  pntsAlpha = 1
)
}
\arguments{
\item{pntsAlpha}{}
}
\description{
Title
}
