% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducePointSet.R
\name{reducePointSet}
\alias{reducePointSet}
\title{Title}
\usage{
reducePointSet(ps, c = 0.85, s = 0.95, e = 0.05)
}
\arguments{
\item{ps}{Set of points as a matrix with x- and y-column}

\item{e}{}
}
\description{
Convenience function that resolves the returned indices of
\link{getPointsToRemove} to points with coordinates
}
