% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRatedDiasForElems.R
\name{getDiagramsRated}
\alias{getDiagramsRated}
\title{Title}
\usage{
getDiagramsRated(inData, categoryCol = "category", nCores = 1)
}
\arguments{
\item{inData}{data frame with one category column and element value columns}

\item{categoryCol}{name of the category column in the \code{inData}}

\item{nCores}{Number of core to use when rating diagrams in parallel}
}
\value{
data frame holding the x- and y-axis and the overlap rate of this diagram
}
\description{
\itemize{
\item df [\code{categoryCol}, el1, el2, ..., elN
\item diagrams wil be generated
\item and all these dias will be rated
}
}
