% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDiagram.R
\name{processDiagram}
\alias{processDiagram}
\title{Title}
\usage{
processDiagram(
  dataSrc,
  xel = "x",
  yel = "y",
  categoryCol = "category",
  scaleAxes = TRUE
)
}
\arguments{
\item{categoryCol}{}
}
\description{
Title
}
